refactor directory struct & add pkg
