+++ pkg/cpu/go.mod
@@ -1,0 +1,12 @@
+module github.com/go-kratos/sra/pkg/cpu
+
+go 1.16
+
+require (
+	github.com/StackExchange/wmi v1.2.1 // indirect
+	github.com/pkg/errors v0.9.1
+	github.com/shirou/gopsutil v3.21.7+incompatible
+	github.com/stretchr/testify v1.7.0
+	github.com/tklauser/go-sysconf v0.3.7 // indirect
+	golang.org/x/sys v0.0.0-20210806184541-e5e7981a1069 // indirect
+)
