+++ pkg/cpu/stat_test.go
