+++ pkg/cpu/utils.go
