+++ pkg/window/counter_test.go
@@ -1,4 +1,4 @@
-package bbr
+package window
 
 import (
 	"testing"
