+++ pkg/window/go.mod
@@ -1,0 +1,5 @@
+module github.com/go-kratos/sra/pkg/window
+
+go 1.16
+
+require github.com/stretchr/testify v1.7.0
