+++ ratelimit/bbr/go.mod
@@ -3,10 +3,11 @@ module github.com/go-kratos/sra/ratelimit/bbr
 go 1.16
 
 require (
-	github.com/StackExchange/wmi v1.2.0 // indirect
-	github.com/go-ole/go-ole v1.2.5 // indirect
-	github.com/pkg/errors v0.9.1
-	github.com/shirou/gopsutil v3.21.6+incompatible
+	github.com/go-kratos/sra/pkg/cpu v0.0.0-00010101000000-000000000000
+	github.com/go-kratos/sra/pkg/window v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
-	github.com/tklauser/go-sysconf v0.3.7 // indirect
 )
+
+replace github.com/go-kratos/sra/pkg/window => ../../pkg/window
+
+replace github.com/go-kratos/sra/pkg/cpu => ../../pkg/cpu
